Use SplObjectStorage as an Identity Map. This makes the BankAccount::getId() and BankAccount::setId() methods superfluous.

+++ src/persistence/BankAccountMapper.php
@@ -2,10 +2,12 @@
 class BankAccountMapper
 {
     protected $db;
+    protected $identityMap;
 
     public function __construct(PDO $db)
     {
-        $this->db = $db;
+        $this->db          = $db;
+        $this->identityMap = new SplObjectStorage;
     }
 
     public function getAllIds()
@@ -21,6 +23,16 @@ class BankAccountMapper
 
     public function findById($id)
     {
+        $this->identityMap->rewind();
+
+        while ($this->identityMap->valid()) {
+            if ($this->identityMap->getInfo() == $id) {
+                return $this->identityMap->current();
+            }
+
+            $this->identityMap->next();
+        }
+
         $result = $this->db->query(
           sprintf(
             'SELECT balance FROM bankaccount WHERE id = %d;',
@@ -38,7 +50,8 @@ class BankAccountMapper
 
         $ba = new BankAccount;
         $ba->setBalance($balance);
-        $ba->setId($id);
+
+        $this->identityMap[$ba] = $id;
 
         return $ba;
     }
@@ -52,27 +65,37 @@ class BankAccountMapper
           )
         );
 
-        $ba->setId((int)$this->db->lastInsertId());
+        $this->identityMap[$ba] = (int)$this->db->lastInsertId();
     }
 
     public function update(BankAccount $ba)
     {
+        if (!isset($this->identityMap[$ba])) {
+            throw new MapperException('Object has no ID, cannot update.');
+        }
+
         $this->db->exec(
           sprintf(
             'UPDATE bankaccount SET balance = %f WHERE id = %d;',
             $ba->getBalance(),
-            $ba->getId()
+            $this->identityMap[$ba]
           )
         );
     }
 
     public function delete(BankAccount $ba)
     {
+        if (!isset($this->identityMap[$ba])) {
+            throw new MapperException('Object has no ID, cannot delete.');
+        }
+
         $this->db->exec(
           sprintf(
             'DELETE FROM bankaccount WHERE id = %d;',
-            $ba->getId()
+            $this->identityMap[$ba]
           )
         );
+
+        unset($this->identityMap[$ba]);
     }
 }
